Rename addDiscount.ts to discount.ts and update imports
Renamed modules/zuora/src/addDiscount.ts to modules/zuora/src/discount.ts for clarity and consistency. Updated all relevant imports in discountEndpoint.ts and associated test files to reference the new module name.
